Fixed the templates and static css file
Fixed the templates,
Updated the index.html in how it looks, made it took look similar to that of previous index.html and managed to fix template issues.
Rearranged post forms in better position then it was.
updated style.css for our project part 2 and the new index.html
like everything chat -> converted to posts.