Changes to board displaying from server, css, html
Just some changes to Board html and css.
For now user is directed to a gameboard (Finally shows board) /board/<lobby_id> only if its a user that didnt create the lobby and presses on join-lobby. Might have to redirect to a board after user creates, this way works for both non created by user and created by user as well.